Initial commit of easage
easage is MIT-licensed software written by Taryn Hill <taryn@phrohdoh.com>

This is a work-in-progress re-write of my `libbig` project
which is available at <https://github.com/Phrohdoh/libbig>.

This project is *feature incomplete* compared to libbig.

The largest change in this project is that archives are memory-mapped
to avoid loading all of the data into memory up-front.

I have not written benchmarks to prove whether or not this is a
net-positive (or even -neutral) change compared to eager loading.