Fix a bug introduced by 11d2c2f4569322887de59d3f88a62617dab061e5
This commit closes #19 on the GitHub issue tracker.
ref: #19

The referenced commit introduced the '--strip-prefix' parameter to easage-pack.
Previously we calculated the size of the metadata table and of the entire
archive using the source file path (on the system performing the pack)
when what we meant to do was use the entry name that would be written
into the archive.