Rewrite archive 'packing' to be more flexible
This commit closes #41 on the GitHub issue tracker.
ref: https://github.com/Phrohdoh/easage/issues/41

Introduce a `packer::pack` function which takes a vec of `(name, data)` tuples
and an `easage::Kind` and returns a `Result<Archive>`.

Assuming a `pack` call returns `Ok::<Archive, _>(a)` then `a`'s data can be acquired
via `a.as_slice()` which can be used to, for example, write to disk via `f.write_all()`.

Rewrite `packer::pack_directory` to call `packer::pack`.